+++ libcircuit_reliability/src/simulations_handling/spectre_simulations/critical_parameter_nd_parameters_sweep_simulation.cpp
@@ -727,7 +727,6 @@ int CriticalParameterNDParameterSweepSimulation::GnuplotPlane(
 	}else{
 		plane.set_general_image_path( outputImagePath );
 	}
-
 	return std::system( execCommand.c_str() );
 }